# docs/API.yaml
openapi: 3.0.3
info:
  title: PCParts API
  description: Find the best parts at the best price
  version: 0.1.0

servers:
  - url: http://localhost:8000/api/v1
    description: Local development

paths:
  /parts:
    get:
      summary: List parts
      parameters:
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/Category'
        - name: q
          in: query
          description: Search query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of parts
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Part'
                  total:
                    type: integer

  /parts/{id}:
    get:
      summary: Get part by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Part details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartDetail'
        '404':
          description: Part not found

  /parts/{id}/prices:
    get:
      summary: Get current prices for a part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Prices from all retailers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Price'

  /parts/{id}/history:
    get:
      summary: Get price history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        '200':
          description: Price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistory'

  /categories:
    get:
      summary: List categories
      responses:
        '200':
          description: Available categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'

  /retailers:
    get:
      summary: List retailers
      responses:
        '200':
          description: Supported retailers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Retailer'

  /builds:
    post:
      summary: Create a new build
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildCreate'
      responses:
        '201':
          description: Build created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Build'

  /builds/{id}:
    get:
      summary: Get build by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Build details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildDetail'
    put:
      summary: Update a build
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildUpdate'
      responses:
        '200':
          description: Build updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Build'

  /builds/{id}/compatibility:
    get:
      summary: Check build compatibility
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Compatibility check result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompatibilityResult'

  /search:
    get:
      summary: Full-text search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: category
          in: query
          schema:
            $ref: '#/components/schemas/Category'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Part'
                  total:
                    type: integer

components:
  schemas:
    Category:
      type: string
      enum:
        - cpu
        - gpu
        - motherboard
        - ram
        - storage
        - psu
        - case
        - cooler

    Part:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        brand:
          type: string
        category:
          $ref: '#/components/schemas/Category'
        lowest_price:
          type: number
          nullable: true
        image_url:
          type: string
          nullable: true

    PartDetail:
      allOf:
        - $ref: '#/components/schemas/Part'
        - type: object
          properties:
            specs:
              type: object
              additionalProperties: true
            prices:
              type: array
              items:
                $ref: '#/components/schemas/Price'

    Price:
      type: object
      properties:
        retailer_id:
          type: string
        retailer_name:
          type: string
        price:
          type: number
        currency:
          type: string
          default: USD
        url:
          type: string
        in_stock:
          type: boolean
        scraped_at:
          type: string
          format: date-time

    PriceHistory:
      type: object
      properties:
        retailer_id:
          type: string
        date:
          type: string
          format: date
        price:
          type: number

    Retailer:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        url:
          type: string
        logo_url:
          type: string
          nullable: true

    BuildCreate:
      type: object
      properties:
        name:
          type: string

    BuildUpdate:
      type: object
      properties:
        name:
          type: string
        part_ids:
          type: array
          items:
            type: string

    Build:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    BuildDetail:
      allOf:
        - $ref: '#/components/schemas/Build'
        - type: object
          properties:
            parts:
              type: array
              items:
                $ref: '#/components/schemas/PartDetail'
            total_price:
              type: number
            compatibility:
              $ref: '#/components/schemas/CompatibilityResult'

    CompatibilityResult:
      type: object
      properties:
        compatible:
          type: boolean
        issues:
          type: array
          items:
            $ref: '#/components/schemas/CompatibilityIssue'

    CompatibilityIssue:
      type: object
      properties:
        severity:
          type: string
          enum: [error, warning, info]
        message:
          type: string
        parts:
          type: array
          items:
            type: string
          description: IDs of parts involved